Remove languager server wrapper interface
