+++ src/main/java/com/github/lsp4intellij/extensions/LSPExtensionManager.java
@@ -4,7 +4,6 @@ import com.github.lsp4intellij.client.languageserver.ServerOptions;
 import com.github.lsp4intellij.client.languageserver.requestmanager.DefaultRequestManager;
 import com.github.lsp4intellij.client.languageserver.requestmanager.RequestManager;
 import com.github.lsp4intellij.client.languageserver.wrapper.LanguageServerWrapper;
-import com.github.lsp4intellij.client.languageserver.wrapper.LanguageServerWrapperImpl;
 import com.github.lsp4intellij.editor.EditorEventManager;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.editor.event.DocumentListener;
@@ -18,7 +17,7 @@ public interface LSPExtensionManager {
             LanguageServer server, LanguageClient client, ServerCapabilities serverCapabilities);
 
     <T extends EditorEventManager> T getExtendedEditorEventManagerFor(Editor editor, DocumentListener documentListener,
-            RequestManager requestManager, ServerOptions serverOptions, LanguageServerWrapperImpl wrapper);
+            RequestManager requestManager, ServerOptions serverOptions, LanguageServerWrapper wrapper);
 
     Class<? extends LanguageServer> getExtendedServerInterface();
